+++ Statistics/src/main/java/com/coviam/quizMedia/Statistics/repository/LeaderBoardRepository.java
@@ -1,7 +1,9 @@
 package com.coviam.quizMedia.Statistics.repository;
 
+import com.coviam.quizMedia.Statistics.entity.LeaderBoardEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LeaderBoardRepository {
+public interface LeaderBoardRepository extends MongoRepository<LeaderBoardEntity,String> {
 }
