+++ Statistics/src/main/java/com/coviam/quizMedia/Statistics/services/impl/StateServiceImpl.java
@@ -1,7 +1,8 @@
-package com.coviam.quizMedia.Statistics.services;
+package com.coviam.quizMedia.Statistics.services.impl;
 
 import com.coviam.quizMedia.Statistics.entity.State;
 import com.coviam.quizMedia.Statistics.repository.StateRepository;
+import com.coviam.quizMedia.Statistics.services.StateService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
